Update default.html
This is a link to my codepen just in chase you are wondering why it seem like I copy and paste. I did, but it was my work.
website: https://codepen.io/michael897_/pen/MGrQZP?editors=1000
Sidenote: I was thinking of a mini project or maybe a whole project, depends on how you think of it, about the NBA. I was thinking about making a trivia based on the NBA. It was just a thought and maybe you can give me feed back on what you think. It would questions and the user would have to answer and there would be points and stuff along those lines.